+++ NEWS.md
@@ -5,6 +5,9 @@
 * Added "exact" mode to ReadInfo() to read info.txt files with exact info.txt file structure
 * Added AddInfoLine(), RemoveInfoLine(), and WriteInfo() functions to help with info.txt file manipulation
 * Added AllSimToPar() and BestSimsToPar() functions to update a par.txt file with parameter values from an allsim.txt or bestsims.txt file
+* Updated CompareFiles() to include options for Simass and Subass files
+* Add error message to CompareFiles() to handle if compare.order = FALSE and by = NULL.
+* Update ReadSubass() to be able to read Subass files with zero rows
 
 ## HYPEtools 1.1.0
 *New functions and bug fixes*